+++ app/components/Analyst/CBC/CbcHeader.tsx
@@ -116,8 +116,9 @@ const CbcHeader: React.FC<Props> = ({ query, isFormEditMode }) => {
                 name: 'conditionally_approved',
                 id: 1,
               },
-              { description: 'Reporting Complete', name: 'complete', id: 2 },
               { description: 'Agreement Signed', name: 'approved', id: 3 },
+              { description: 'Reporting Complete', name: 'complete', id: 2 },
+              { description: 'Withdrawn', name: 'withdrawn', id: 4 },
             ]}
             isFormEditMode={isFormEditMode}
           />